Move `DatabaseSeedTask` and test to new `service.seeding` directory, rename `DatabaseSeedController` (#5226)
