+++ server/app/services/seeding/DatabaseSeedTask.java
@@ -1,4 +1,4 @@
-package controllers.dev.seeding;
+package services.seeding;
 
 import static com.google.common.base.Preconditions.checkNotNull;
