+++ server/conf/routes
@@ -199,10 +199,10 @@ GET     /logout                      @org.pac4j.play.LogoutController.logout(req
 
 ## Methods for development
 # Seed the database with test content to develop against, and clear the database
-GET     /dev/seed                        controllers.dev.seeding.DatabaseSeedController.index(request: Request)
-POST    /dev/seedQuestions               controllers.dev.seeding.DatabaseSeedController.seedQuestions()
-POST    /dev/seedPrograms                controllers.dev.seeding.DatabaseSeedController.seedPrograms()
-POST    /dev/seed/clear                  controllers.dev.seeding.DatabaseSeedController.clear()
+GET     /dev/seed                        controllers.dev.seeding.DevDatabaseSeedController.index(request: Request)
+POST    /dev/seedQuestions               controllers.dev.seeding.DevDatabaseSeedController.seedQuestions()
+POST    /dev/seedPrograms                controllers.dev.seeding.DevDatabaseSeedController.seedPrograms()
+POST    /dev/seed/clear                  controllers.dev.seeding.DevDatabaseSeedController.clear()
 GET     /dev/icons                       controllers.dev.IconsController.index(request: Request)
 
 # Change feature flags.
